-r or --recent option added.

+++ review-cctv-monitorings.sh
@@ -1,8 +1,16 @@
 #!/bin/sh
 
-usage="Usage: $(basename $0): [-nf|--no-fetch] [-na|--no-autoplay]: allows to (possibly) fetch from server and review conveniently any set of CCTV recordings dating back from yesterday and the three days before, from the current directory.
-  Without fetching, CCTV recordings are expected to be already available in the current directory (see the fetch-cctv-monitorings.sh script for that, possibly installed in a crontab).
-  With autoplay, recordings are displayed in a row, and offered to be deleted as a whole afterwards (locally and/or on the server). Without autoplay, they are displayed one by one, the user being asked what to do with each of them in turn."
+usage="Usage: $(basename $0): [-r|--recent] [-nf|--no-fetch] [-na|--no-autoplay]: allows to (possibly) fetch from server and review conveniently any set of CCTV recordings dating back from yesterday and the three days before, from the current directory.
+
+Options:
+  -r or --recent: selects only recent recordings, from yesterday's evening, and until now
+  -nf or --no-fetch: without fetching, CCTV recordings are expected to be already available in the current directory (see the fetch-cctv-monitorings.sh script for that, possibly installed in a crontab)
+  -na or --no-autoplay: without autoplay, recordings are displayed one by one, the user being asked what to do with each of them in turn; with autoplay, recordings are displayed in a row, and offered to be deleted as a whole afterwards (locally and/or on the server)
+"
+
+
+select_recent=1
+
 
 # Enabled by default:
 do_fetch=0
@@ -49,6 +57,13 @@ while [ $token_eaten -eq 0 ]; do
 	#[ $verbose -eq 1 ] || echo "Args: $*"
 
 	token_eaten=1
+
+	if [ "$1" = "-r" ] || [ "$1" = "--recent" ]; then
+		echo "(recent selection mode enabled)"
+		shift
+		select_recent=0
+		token_eaten=0
+	fi
 
 	if [ "$1" = "-na" ] || [ "$1" = "--no-autoplay" ]; then
 		echo "(autoplay disabled)"
@@ -142,14 +157,27 @@ if [ $do_fetch -eq 0 ]; then
 
 	cd "${review_dir}"
 
-	yesterday="$(date -d '-1 day' '+%Y%m%d')"
-	day_minus_two="$(date -d '-2 day' '+%Y%m%d')"
-	day_minus_three="$(date -d '-3 day' '+%Y%m%d')"
-	day_minus_four="$(date -d '-4 day' '+%Y%m%d')"
+	if [ $select_recent -eq 1 ]; then
 
-	echo "Fetching as user ${USER} CCTV recordings from ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH} for yesterday (i.e. ${yesterday}) and the three days before (i.e. ${day_minus_two}, ${day_minus_three} and ${day_minus_four}):"
+		yesterday="$(date -d '-1 day' '+%Y%m%d')"
+		day_minus_two="$(date -d '-2 day' '+%Y%m%d')"
+		day_minus_three="$(date -d '-3 day' '+%Y%m%d')"
+		day_minus_four="$(date -d '-4 day' '+%Y%m%d')"
 
-	${scp} ${scp_opt} ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_four}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_three}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_two}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${yesterday}*.mkv . 2>/dev/null
+		echo "Fetching as user ${USER} CCTV recordings from ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH} for yesterday (i.e. ${yesterday}) and the three days before (i.e. ${day_minus_two}, ${day_minus_three} and ${day_minus_four}):"
+
+		${scp} ${scp_opt} ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_four}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_three}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_two}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${yesterday}*.mkv . 2>/dev/null
+
+	else
+
+		yesterday="$(date -d '-1 day' '+%Y%m%d')"
+		today="$(date '+%Y%m%d')"
+
+		echo "Fetching as user ${USER} CCTV recordings from ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH} for yesterday (i.e. ${yesterday}) and today (i.e. ${today}):"
+
+		${scp} ${scp_opt} ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${yesterday}*.mkv ${CCTV_USER}@${CCTV_SERVER}:${CCTV_BASE_PATH}/${CCTV_PREFIX}*${today}*.mkv . 2>/dev/null
+
+	fi
 
 	# Disabled, as returns an error as soon as there is no recording for at
 	# least one of these days (which is however normal):
@@ -392,17 +420,29 @@ if [ $do_fetch -eq 0 ] && [ ${auto_play} -eq 1 ] && [ -n "${recordings}" ]; then
 			ssh_opt="-p ${SSH_PORT}"
 		fi
 
-		ssh=$(which ssh)
+		ssh="$(which ssh)"
 
-		if ${ssh} ${ssh_opt} ${CCTV_USER}@${CCTV_SERVER} /bin/rm -f ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_three}*.mkv ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_two}*.mkv ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${yesterday}*.mkv; then
+		if [ $select_recent -eq 1 ]; then
 
-			echo "Deleted!"
+			if ${ssh} ${ssh_opt} ${CCTV_USER}@${CCTV_SERVER} /bin/rm -f ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_three}*.mkv ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${day_minus_two}*.mkv ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${yesterday}*.mkv; then
+
+				echo "Deleted!"
+
+			fi
+
+		else
+
+			if ${scp} ${scp_opt} ${CCTV_USER}@${CCTV_SERVER} /bin/rm -f ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${yesterday}*.mkv ${CCTV_BASE_PATH}/${CCTV_PREFIX}*${today}*.mkv;  then
+
+				echo "Deleted!"
+
+			fi
 
 		fi
 
 	else
 
-		echo "No remote recordings deleted."
+			echo "No remote recordings deleted."
 
 	fi
